refactor(researcher): Mover SQLiteResearchRepository a infraestructura
- Movido SQLiteResearchRepository a infrastructure/repositories/
- Creada clase ResearchQueries para centralizar consultas SQL
- Mejorada organización de queries en archivo independiente
- Actualizadas importaciones en researcher.py
