+++ app/agents/researcher/infrastructure/repositories/__init__.py
@@ -1,0 +1,3 @@
+from .sqlite_repository import SQLiteResearchRepository
+
+__all__ = ['SQLiteResearchRepository']
